Added support for creating two links.
Pulled in jQuery and added a small script to handle when the user
enters both text and a url. Now, when this happens, the user will
be alerted and a radio box will appear. If the user clicks the box,
two links (with the same title) will be created, one to the url
and one to the text. Otherwise, only the link to the text will be
created.

Also added swap and patch files to the gitignore.