Refactored the javascript and added more functionality.
First, made it so that when the text in one of either the text or url fields
is deleted, the 'Create Both Links' button will fade out. Also fixed it
so that the button will only fadeIn after the user has been warned and
there is text in both the url and text fields.

Wrote a URL validator function and added a submit button handler
to call the validator on mouseover.

Also added a notifier to the links list that indicates if the particular
link is internal or external.

Also added documentation and comments and started updating some of the
unit tests.