Cleaned up some documentation, fixed a minor bug, increased error checking in UserText(), and refactored how sripts.js handles error messages.
Commit 8958923d6b5243a1d9c671ad050373349636bd75 changed the behavior of
LinkReadOnly::getLinkById() to return an instance of
Application\Domain\Entiy\Link instead of the AtlasORM Record object. This
brought to light a small bug in the Atlas query statement in that it was
not fetching the related User record with the link db record. As such,
when getLinkById() tried to create a new Link instance from the fetched
db record, PHP 'Trying to get property of non-object' notices started
showing up in the error log, since the related User record was not being
fetched, as well. Now, the related User record is being fetched.

Also, strengthened the error checking of the UserText class. Now, the
value of the 'id' parameter from the query string must match a decimal
digit instead of merely being defined.

Moved all of the error message storage and handling in scripts.js into
an object in order to clean up the code and make it more extensible for
future messages.